# add common con support 3x3s12,5x5s12,7x7s12, fix leaky_relu bug, optimize direct con by exchange loop