--- a/lite/backends/x86/math/conv_direct_fp32.cc
+++ b/lite/backends/x86/math/conv_direct_fp32.cc
@@ -22,16 +22,16 @@ limitations under the License. */
 #else
 #include <emmintrin.h>
 #endif
-#include "lite/backends/x86/math/conv3x3s2_direct_fp32.h"
+#include "lite/backends/x86/math/conv_direct_fp32.h"
 
 namespace paddle {
 namespace lite {
 namespace x86 {
 namespace math {
 
-#define GET_OFF(field) offsetof(jit_3x3s2param, field)
+#define GET_OFF(field) offsetof(jit_param, field)
 
-struct jit_3x3s2param {
+struct jit_param {
   const float* in_row_addr;
   const float* kernel_addr;
   float* out_row_addr;
@@ -40,9 +40,9 @@ struct jit_3x3s2param {
   size_t wh;
 };
 
-conv_direct_3x3s2::conv_direct_3x3s2() : JitCode(8192, Xbyak::AutoGrow) {}
+conv_direct::conv_direct() : JitCode(8192, Xbyak::AutoGrow) {}
 
-void conv_direct_3x3s2::generate_code(int ic,
+void conv_direct::generate_code(int ic,
                                       int ih,
                                       int iw,
                                       int oc,
@@ -50,16 +50,16 @@ void conv_direct_3x3s2::generate_code(int ic,
                                       int oh,
                                       int ow,
                                       int ph,
-                                      int pw) {
+                                      int pw, 
+                                      int ww,
+                                      int wh,
+                                      int stridew) {
 #ifdef __AVX__
   constexpr int BLOCK = 8;
 #else
   constexpr int BLOCK = 4;
 #endif
 
-  constexpr int ww = 3;
-  constexpr int wh = 3;
-  constexpr int stridew = 2;
   constexpr int oc_block = 4 * BLOCK;
   const int oc_loop_n =
       oc_expand / oc_block;  // deal with every 32 output channels
@@ -73,19 +73,17 @@ void conv_direct_3x3s2::generate_code(int ic,
 
   int new_ow = 0;  // 0 or 3 or 6....
   bool right = false;
-  if (ph == 0 && pw == 0) {
+  if (pw == 0) {
     new_ow = ow / ow_bulk * ow_bulk;
     new_ow -= ow_bulk;
-  } else if (ph == 1 && pw == 1) {
+  } else if (pw > 0) {
     // whether padding is required for the rightet output
-    right = (-pw + stridew * (ow - 1) + ww - 1) == ((iw - 1) + pw);
+    right = stridew * (ow - 1) + ww - iw - pw > 0;
     if (right)
       new_ow = ow % ow_bulk ? ow / ow_bulk * ow_bulk : ow - ow_bulk;
     else
       new_ow = ow / ow_bulk * ow_bulk;
     new_ow -= ow_bulk;
-  } else {
-    LOG(FATAL) << "[X86] conv_direct only support 3x3s2 with padding = 0 or 1";
   }
 
   // The number of channels of convolution kernel
@@ -142,6 +140,7 @@ void conv_direct_3x3s2::generate_code(int ic,
   };
 
   // one_line means we only compute one line kernel and one line input
+  // every input row consists of [l_pad's data, real data, r_pad's data]
   auto fmadd_one_line = [=, &temp](
       int oc_group, int ic_group, int l_pad, int r_pad, int bulk) {
     // kernel has ww columns, but I only take one number at a time
@@ -151,7 +150,7 @@ void conv_direct_3x3s2::generate_code(int ic,
         for (int j = 0; j < bulk; j++) {
           // no need to fetch this input
           if (ww_i + stridew * j < l_pad) continue;
-          if (ww_i + stridew * j >= stridew * bulk + 1 - r_pad) continue;
+          if (ww_i + stridew * j >= stridew * (bulk - 1) + ww - r_pad) continue;
 
           Vmm input = Vmm(oc_group / BLOCK * bulk + j);
           temp = (ww_i - l_pad + stridew * j + ic_i * ihw) * sizeof(float);
@@ -167,7 +166,7 @@ void conv_direct_3x3s2::generate_code(int ic,
           for (int j = 0; j < bulk; j++) {
             // no need to fetch this input
             if (ww_i + stridew * j < l_pad) continue;
-            if (ww_i + stridew * j >= stridew * bulk + 1 - r_pad) continue;
+            if (ww_i + stridew * j >= stridew * (bulk - 1) + ww - r_pad) continue;
 
             Vmm input = Vmm(oc_group / BLOCK * bulk + j);
             Vmm res(oc_gi / BLOCK * bulk + j);
@@ -178,9 +177,12 @@ void conv_direct_3x3s2::generate_code(int ic,
     }
   };
 
+  // compute output whole line
+  // three parts: [left, middle, right]
   auto cal_out_whole_line = [=, &temp](int oc_group, int ic_group) {
     int ow_bulk_i = ow / ow_bulk;
 
+    
     auto cal_bulk = [=, &temp](
         int oc_group, int ic_group, int l_pad, int r_pad, int bulk) {
       load(oc_group, bulk);
@@ -207,17 +209,26 @@ void conv_direct_3x3s2::generate_code(int ic,
 
     // entry !
     // left
+    // we need check if there is a left part
+    // besided, we need check if the first ow_bulk output is 
+    // associated with right boundry
+    int ow_rpad = 0;
+    int first_ow_bulk_rpad = (ow_bulk - 1) * stridew + ww - pw - iw;
+    if (first_ow_bulk_rpad > 0)// means ow <= 3
+      ow_rpad = (ow - 1) * stridew + ww - pw - iw;
+ 
     mov(ow_bulk_i_xb, 0);
-    if (pw == 1 || new_ow < 0) {
+    if (pw > 0 || ow <= ow_bulk) {
       int temp_rpad = 0;
-      if (pw == 1 && right && new_ow < 0) temp_rpad = pw;
-      cal_bulk(oc_group, ic_group, pw, temp_rpad, new_ow < 0 ? ow : ow_bulk);
-      inc(ow_bulk_i_xb);
-      add(in_row_addr_xb, 5 * sizeof(float));
-      add(out_row_addr_xb, ow_bulk * BLOCK * sizeof(float));
+      if (ow_rpad > 0) temp_rpad = ow_rpad;
+      cal_bulk(oc_group, ic_group, pw, temp_rpad, ow <= ow_bulk ? ow : ow_bulk);
 
+      inc(ow_bulk_i_xb);
+      add(in_row_addr_xb, (stridew * ow_bulk - pw) * sizeof(float));
+      add(out_row_addr_xb, ow_bulk * BLOCK * sizeof(float));
       ow_bulk_i--;
     }
+
     // judge whether there is an middle part
     if (ow_bulk_i > 0 && !(right && ow == 2 * ow_bulk)) {
       // middle !
@@ -237,7 +248,8 @@ void conv_direct_3x3s2::generate_code(int ic,
     // ow_remain = rightest index - ??
     int ow_remain = (ow - 1) - (new_ow + 2);
     if (ow_remain > 0 && new_ow >= 0) {  // right exists
-      int r_pad = right ? pw : 0;
+      ow_rpad = (ow - 1) * stridew + ww - pw - iw;
+      int r_pad = right ? ow_rpad : 0;
       cal_bulk(oc_group, ic_group, 0, r_pad, ow_remain);
     }
   };
@@ -289,7 +301,7 @@ void conv_direct_3x3s2::generate_code(int ic,
   postCode();
 }
 
-void conv_direct_3x3s2::run(const float* i_data,
+void conv_direct::run(const float* i_data,
                             const float* trans_weight,
                             float* trans_out,
                             int bs,
@@ -301,16 +313,16 @@ void conv_direct_3x3s2::run(const float* i_data,
                             int oh,
                             int ow,
                             int ph,
-                            int pw) {
+                            int pw,
+                            int wh,
+                            int ww,
+                            int strideh) {
 #ifdef __AVX__
   constexpr int BLOCK = 8;
 #else
   constexpr int BLOCK = 4;
 #endif
 
-  constexpr int ww = 3;
-  constexpr int wh = 3;
-  constexpr int strideh = 2;
   constexpr int oc_block = 4 * BLOCK;
   constexpr int ic_block = 8;
   int wc = ic;
@@ -329,7 +341,7 @@ void conv_direct_3x3s2::run(const float* i_data,
                             int wh) {
       for (int ic_i = 0; ic_i < ic; ic_i += ic_block) {
         for (int oc_gi = 0; oc_gi < oc; oc_gi += oc_block) {
-          jit_3x3s2param param;
+          jit_param param;
           param.in_row_addr = in_row_addr + ic_i * ihw;
           param.kernel_addr =
               trans_weight + oc_gi / BLOCK * whwB * wc + ic_i * whwB;
@@ -338,8 +350,8 @@ void conv_direct_3x3s2::run(const float* i_data,
           param.ic = ic_i + ic_block - 1 < ic ? ic_block : ic - ic_i;
           param.wh = wh;
 
-          void (*f)(jit_3x3s2param*) =
-              CodeGenerator::getCode<void (*)(jit_3x3s2param*)>();
+          void (*f)(jit_param*) =
+              CodeGenerator::getCode<void (*)(jit_param*)>();
           f(&param);
         }
       }
@@ -349,12 +361,16 @@ void conv_direct_3x3s2::run(const float* i_data,
     float* out_row_addr = trans_out + bs_i * ochw;
 
     int oh_i = 0;
-    if (ph == 1) {  // upper boundry
-      int temp_wh = ih >= 2 ? 2 : 1;
+    if (ph > 0) {  // upper boundry
+      int temp_wh = wh - ph; // we olny need deal with temp_wh rows not wh rows!
+
+      // check if the kernel will occupy lower boundry
+      // if so, we need decrease temp_wh again 
+      if (ih + ph < wh) temp_wh -= (wh - ih - ph);
       cal_out_line(
-          in_row_addr, trans_weight + ww * BLOCK, out_row_addr, temp_wh);
+          in_row_addr, trans_weight + ph * ww * BLOCK, out_row_addr, temp_wh);
       oh_i++;
-      in_row_addr += iw;
+      in_row_addr += (strideh - ph) * iw;
       out_row_addr += ow * BLOCK;
     }
     // middle
@@ -364,13 +380,14 @@ void conv_direct_3x3s2::run(const float* i_data,
       in_row_addr += strideh * iw;
     }
 
-    // lower boundary
     if (oh_i >= oh) continue;
+    
+    // lower boundary, 
+    // compute how many boundry rows is used to the lowerest output row
+    int lower =  strideh * (oh - 1) + wh - ph - ih;
 
-    bool lower = (-ph + strideh * (oh - 1) + wh - 1) == ((ih - 1) + ph);
-
-    if (ph == 1 && lower) {
-      cal_out_line(in_row_addr, trans_weight, out_row_addr, 2);
+    if (lower > 0) {
+      cal_out_line(in_row_addr, trans_weight, out_row_addr, wh - lower);
     } else {
       cal_out_line(in_row_addr, trans_weight, out_row_addr, wh);
     }
@@ -379,7 +396,7 @@ void conv_direct_3x3s2::run(const float* i_data,
 
 // we always assume oc % BLOCK == 0!
 // convert [N C/8 H W 8] to [N C H W]!
-void conv_direct_3x3s2_tranpose_out(int bs,
+void conv_direct_transpose_out(int bs,
                                     int oc,
                                     int oh,
                                     int ow,