--- a/lite/kernels/x86/conv_direct.cc
+++ b/lite/kernels/x86/conv_direct.cc
@@ -24,9 +24,6 @@ namespace x86 {
 template <>
 void DirectConv<PRECISION(kFloat), PRECISION(kFloat)>::Run() {
   auto& param = this->Param<param_t>();
-  // CHECK_EQ(param.strides[0], 2);
-  // CHECK_EQ(param.strides[1], 2);
-  // auto& ctx = this->ctx_->template As<X86Context>();
 
   const auto* i_data = param.x->data<float>();
   const auto* b_data = param.bias ? param.bias->data<float>() : nullptr;