… lite/kernels/x86/conv_compute.cc
… lite/kernels/x86/conv_direct.h