## 3x3s1p1

| 输入               | old  | new  | 时间减少 |
| ------------------ | ---- | ---- | -------- |
| 1,64,56,56：64核   | 2.3  | 1.7  | 26%      |
| 1,128,28,28：128核 | 2.3  | 1.64 | 28%      |
| 1,256,14,14：256核 | 2.36 | 1.6  | 28%      |
| 1,512,7,7：512核   | 3.13 | 1.7  | 45%      |
| 1,16,448,448：16核 | 50   | 10   | 80%      |


## 5x5s2p2

| 输入              | old     | new      | 时间减少 |
| ----------------- | ------- | -------- | -------- |
| 1,3,224,224：32核 | 0.93 ms | 0.516 ms | 44%      |


## 7x7s2p3

- 但其实我测试的是7x7s2p2

| 输入              | old | new | 时间减少 |
| ----------------- | --- | --- | -------- |
| 1,3,224,224：32核 | 1.7 | 0.9 | 47%      |

## resnet

| 模型        | old | new | 时间减少 |
| ----------- | --- | --- | -------- |
| resnet50    | 90  | 80  | 11%      |
| resnet50_vd | 98  | 84  | 14%      |